sentinelayer-api/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
sqlalchemy>=2.0
asyncpg>=0.29
pydantic>=2.6
pydantic-settings>=2.2
python-jose[cryptography]>=3.3
httpx>=0.27
redis>=5.0
alembic>=1.13
pytest>=8.0
pytest-asyncio>=0.23
